Refactor dump download code and support local listen dumps
