+++ listenbrainz/spark/request_queries.json
@@ -57,22 +57,22 @@
   "import.dump.full_newest": {
     "name": "import.dump.full_newest",
     "description": "Import the newest available full dump into the spark cluster",
-    "params": []
+    "params": ["local"]
   },
   "import.dump.full_id": {
     "name": "import.dump.full_id",
     "description": "Import full dump with the specified ID into the spark cluster",
-    "params": ["dump_id"]
+    "params": ["dump_id", "local"]
   },
   "import.dump.incremental_newest": {
     "name": "import.dump.incremental_newest",
     "description": "Import the newest available incremental dump into the spark cluster",
-    "params": []
+    "params": ["local"]
   },
   "import.dump.incremental_id": {
     "name": "import.dump.incremental_id",
     "description": "Import incremental dump with the specified ID into the spark cluster",
-    "params": ["dump_id"]
+    "params": ["dump_id", "local"]
   },
   "import.dump.mlhd": {
     "name": "import.dump.mlhd",
