+++ listenbrainz_spark/ftp/download.py
@@ -1,11 +1,12 @@
+import logging
 import os
 import tempfile
 import time
-import logging
 
 from listenbrainz_spark import config
-from listenbrainz_spark.ftp import ListenBrainzFTPDownloader, DumpType, ListensDump
+from listenbrainz_spark.dump import DumpType, ListensDump
 from listenbrainz_spark.exceptions import DumpNotFoundException
+from listenbrainz_spark.ftp import ListenBrainzFTPDownloader
 
 # mbid_msid_mapping_with_matchable is used.
 # refer to: http://ftp.musicbrainz.org/pub/musicbrainz/listenbrainz/labs/mappings/
@@ -19,38 +20,6 @@ logger = logging.getLogger(__name__)
 
 class ListenbrainzDataDownloader(ListenBrainzFTPDownloader):
 
-    def get_dump_name_to_download(self, dump, dump_id, dump_id_pos):
-        """ Get name of the dump to be downloaded.
-
-            Args:
-                dump (list): Contents of the directory from which dump will be downloaded.
-                dump_id (int): Unique indentifier of dump to be downloaded .
-                dump_id_pos (int): Unique identifier position in dump name.
-
-            Returns:
-                req_dump (str): Name of the dump to be downloaded.
-        """
-        if dump_id:
-            req_dump = None
-            for dump_name in dump:
-                if int(dump_name.split('-')[dump_id_pos]) == dump_id:
-                    req_dump = dump_name
-                    break
-            if req_dump is None:
-                err_msg = "Could not find dump with ID: {}. Aborting...".format(dump_id)
-                raise DumpNotFoundException(err_msg)
-        else:
-            req_dump = dump[-1]
-        return req_dump
-
-    def get_listens_dump_file_name(self, dump_name):
-        """ Get the name of Spark listens dump name archive.
-
-            Returns:
-                str : Spark listens dump archive name.
-        """
-        return ListensDump.from_ftp_dir(dump_name).get_dump_file()
-
     def download_listens(self, directory, listens_dump_id=None, dump_type: DumpType = DumpType.FULL):
         """ Download listens to dir passed as an argument.
 
@@ -63,25 +32,26 @@ class ListenbrainzDataDownloader(ListenBrainzFTPDownloader):
             Returns:
                 dest_path (str): Local path where listens have been downloaded.
                 listens_file_name (str): name of downloaded listens dump.
-                dump_id (int): Unique indentifier of downloaded listens dump.
+                dump_id (int): Unique identifier of downloaded listens dump.
         """
-        if dump_type == DumpType.INCREMENTAL:
-            ftp_cwd = os.path.join(config.FTP_LISTENS_DIR, 'incremental/')
-        else:
-            ftp_cwd = os.path.join(config.FTP_LISTENS_DIR, 'fullexport/')
-        self.connection.cwd(ftp_cwd)
-        listens_dump_list = sorted(self.list_dir(), key=lambda x: int(x.split('-')[2]))
+        dump_directories = self.list_dump_directories(dump_type)
+
+        listens_dump_list = sorted(dump_directories, key=lambda x: int(x.split('-')[2]))
         req_listens_dump = self.get_dump_name_to_download(listens_dump_list, listens_dump_id, 2)
-        dump_id = req_listens_dump.split('-')[2]
+        listens_file_name = self.get_listens_dump_file_name(req_listens_dump)
+        dump_id = int(req_listens_dump.split('-')[2])
 
         self.connection.cwd(req_listens_dump)
-        listens_file_name = self.get_listens_dump_file_name(req_listens_dump)
 
         t0 = time.monotonic()
         logger.info('Downloading {} from FTP...'.format(listens_file_name))
         dest_path = self.download_dump(listens_file_name, directory)
         logger.info('Done. Total time: {:.2f} sec'.format(time.monotonic() - t0))
-        return dest_path, listens_file_name, int(dump_id)
+        self.connection.close()
+        return dest_path, listens_file_name, dump_id
+
+    def load_listens(self, directory, listens_dump_id=None, dump_type: DumpType = DumpType.FULL):
+        return self.download_listens(directory, listens_dump_id, dump_type)
 
     def download_artist_relation(self, directory, artist_relation_dump_id=None):
         """ Download artist relation to dir passed as an argument.
@@ -118,21 +88,10 @@ class ListenbrainzDataDownloader(ListenBrainzFTPDownloader):
                 dump_name = f.readline().strip()
         self.connection.cwd(dump_name)
 
-        logger.info(f"Downloading release.tar.gz of dump {dump_name} from FTP...")
-        t0 = time.monotonic()
         filename = "release.tar.xz"
+        logger.info(f"Downloading {filename} of dump {dump_name} from FTP...")
+        t0 = time.monotonic()
         dest = os.path.join(directory, filename)
         self.download_file_binary(filename, dest)
         logger.info(f"Done. Total time: {time.monotonic() - t0:.2f} sec")
         return dest
-
-    def get_latest_dump_id(self, dump_type: DumpType):
-        if dump_type == DumpType.INCREMENTAL:
-            ftp_cwd = os.path.join(config.FTP_LISTENS_DIR, 'incremental/')
-        else:
-            ftp_cwd = os.path.join(config.FTP_LISTENS_DIR, 'fullexport/')
-        self.connection.cwd(ftp_cwd)
-
-        listens_dumps = [ListensDump.from_ftp_dir(name) for name in self.list_dir()]
-        listens_dumps.sort(key=lambda x: x.dump_id)
-        return listens_dumps[-1].dump_id
