+++ listenbrainz_spark/request_consumer/jobs/import_dump.py
@@ -1,18 +1,16 @@
 """ Spark job that downloads the latest listenbrainz dumps and imports into HDFS
 """
+import logging
 import shutil
 import tempfile
 import time
-import logging
 from datetime import datetime
 
 import listenbrainz_spark.request_consumer.jobs.utils as utils
-from listenbrainz_spark.path import INCREMENTAL_DUMPS_SAVE_PATH
-from listenbrainz_spark.ftp import DumpType
+from listenbrainz_spark.dump import DumpType
+from listenbrainz_spark.dump.local import ListenbrainzLocalDumpLoader
 from listenbrainz_spark.ftp.download import ListenbrainzDataDownloader
 from listenbrainz_spark.hdfs.upload import ListenbrainzDataUploader
-from listenbrainz_spark.request_consumer import request_consumer
-from listenbrainz_spark.utils import read_files_from_HDFS
 
 logger = logging.getLogger(__name__)
 
@@ -33,13 +31,12 @@ def import_full_dump_to_hdfs(dump_id: int = None, local: bool = False) -> str:
         the name of the imported dump
     """
     with tempfile.TemporaryDirectory() as temp_dir:
-        downloader = ListenbrainzDataDownloader()
-        src, dump_name, dump_id = downloader.download_listens(
+        loader = ListenbrainzLocalDumpLoader() if local else ListenbrainzDataDownloader()
+        src, dump_name, dump_id = loader.load_listens(
             directory=temp_dir,
             dump_type=DumpType.FULL,
             listens_dump_id=dump_id
         )
-        downloader.connection.close()
         ListenbrainzDataUploader().upload_new_listens_full_dump(src)
     utils.insert_dump_data(dump_id, DumpType.FULL, datetime.utcnow())
     return dump_name
@@ -59,7 +56,8 @@ def import_incremental_dump_to_hdfs(dump_id: int = None, local: bool = False) ->
         the name of the imported dump
     """
     with tempfile.TemporaryDirectory() as temp_dir:
-        src, dump_name, dump_id = ListenbrainzDataDownloader().download_listens(
+        loader = ListenbrainzLocalDumpLoader() if local else ListenbrainzDataDownloader()
+        src, dump_name, dump_id = loader.load_listens(
             directory=temp_dir,
             dump_type=DumpType.INCREMENTAL,
             listens_dump_id=dump_id
