+++ listenbrainz_spark/request_consumer/jobs/utils.py
@@ -3,16 +3,16 @@ import logging
 from datetime import datetime
 from typing import Optional
 
+from pyspark.sql import Row
+from pyspark.sql.functions import col
+
+from listenbrainz_spark.dump import DumpType
 from listenbrainz_spark.exceptions import PathNotFoundException
-from listenbrainz_spark.ftp import DumpType
+from listenbrainz_spark.hdfs.utils import (delete_dir, path_exists, rename)
 from listenbrainz_spark.path import IMPORT_METADATA
 from listenbrainz_spark.schema import import_metadata_schema
 from listenbrainz_spark.utils import (create_dataframe, read_files_from_HDFS,
                                       save_parquet)
-from listenbrainz_spark.hdfs.utils import (delete_dir, path_exists, rename)
-from pyspark.sql import Row
-from pyspark.sql.functions import col
-
 
 logger = logging.getLogger(__name__)
